+++ src/main/java/cloud/metaapi/sdk/clients/copy_factory/models/CopyFactoryPortfolioMember.java
@@ -24,6 +24,10 @@ public class CopyFactoryPortfolioMember {
      */
     public Double maxTradeRisk;
     /**
+     * Flag indicating that the strategy should be copied in a reverse direction, or {@code null}
+     */
+    public Boolean reverse;
+    /**
      * Setting indicating whether to enable automatic trade correlation reduction, or {@code null}. Possible
      * settings are not specified (disable correlation risk restrictions), by-strategy (limit correlations on
      * strategy level) or by-symbol (limit correlations on symbol level)